Finish Offering controller, view model, repository and views for schedules section. Fixed some modules to get appropiate display in combo boxes, Fixed some bugs in the services (schedules,  periodDate and Enrollment when deleting using cascade style.

+++ ADYC.Service/PeriodDateService.cs
@@ -66,7 +66,10 @@ namespace ADYC.Service
 
         public void RemoveRange(IEnumerable<PeriodDate> periodDates)
         {
-            _periodDateRepository.RemoveRange(periodDates);
+            if (periodDates.Count() > 0)
+            {
+                _periodDateRepository.RemoveRange(periodDates); 
+            }
         }
 
         private int GetTermIdFromPeriodDates(IEnumerable<PeriodDate> periodDates)
